feat: Convert utils, db, model, dao packages from Java to Kotlin
Phase 1-3 of Java to Kotlin migration:
- Setup Kotlin 1.9.25 with AGP 8.0.2 compatibility
- Convert utils package (13 files): Misc, SimpleDate, Logger, etc.
- Convert err package (1 file): HTTPException
- Convert db package (17 files): Room entities and adapters
- Convert model package (10 files): LedgerAccount, LedgerTransaction, etc.
- Convert dao package (11 files): Room DAOs

Key changes:
- Add @JvmStatic, @JvmField for Java interop compatibility
- Use annotationProcessor (not KSP) during migration for Room
- Fix platform declaration clashes with @get:JvmName annotations
- Add Java compatibility methods where needed

Build verified: ./gradlew assembleDebug passes
Total: 52 files converted, ~7000 lines migrated

+++ app/src/main/java/net/ktnx/mobileledger/ui/templates/TemplateDetailsViewModel.java
@@ -87,7 +87,7 @@ public class TemplateDetailsViewModel extends ViewModel {
 
         if (srcList.size() < 1) {
             final TemplateDetailsItem.Header header = TemplateDetailsItem.createHeader();
-            header.setId(0);
+            header.setId(0L);
             newList.add(header);
             changes = true;
         }
@@ -119,7 +119,7 @@ public class TemplateDetailsViewModel extends ViewModel {
         while (newList.size() < 3) {
             final TemplateDetailsItem.AccountRow accountRow =
                     TemplateDetailsItem.createAccountRow();
-            accountRow.setId(genItemId());
+            accountRow.setId((long) genItemId());
             accountRow.setPosition(newList.size());
             newList.add(accountRow);
             changes = true;
@@ -129,7 +129,7 @@ public class TemplateDetailsViewModel extends ViewModel {
         if (!hasEmptyItem) {
             final TemplateDetailsItem.AccountRow accountRow =
                     TemplateDetailsItem.createAccountRow();
-            accountRow.setId(genItemId());
+            accountRow.setId((long) genItemId());
             accountRow.setPosition(newList.size());
             newList.add(accountRow);
             changes = true;
